% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_stage_score.R
\docType{data}
\name{norm.stage.score}
\alias{norm.stage.score}
\title{Normal Lake Stage Envelope thresholds}
\format{
An object of class \code{data.frame} with 365 rows and 14 columns.
}
\source{
RECOVER, 2020, Lake Okeechobee Performance Measure - Lake State
}
\usage{
data(norm.stage.score)
}
\description{
Thresholds for the normal lake stage envelope to approximate corresponding scores for lake stages

Values are represented as Feet NGVD29
}
\keyword{dataset}
