% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_env.R
\name{norm_env}
\alias{norm_env}
\title{Normal Lake Stage Envelope Score}
\usage{
norm_env(stg.data)
}
\arguments{
\item{stg.data}{see details}
}
\value{
Returns a \code{data.frame} of original data and normal stage elevation score.
}
\description{
Normal Lake Stage Envelope Score
}
\details{
The input \code{stg.data} is a \code{data.frame} with columns:
\itemize{
\item{\code{Date} (as a \code{POSIXct} or \code{Date} variable)}
\item{\code{Data.Value} as stage elevation data in feet (NGVD29)}
}
}
\examples{

# Example dataset (not real data)
dat=data.frame(Date=seq(as.Date("2016-01-01"),
as.Date("2016-03-02"),"1 days"),Data.Value=runif(62,12,18))

norm_env(dat)
}
