% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_env.R
\name{rec_env}
\alias{rec_env}
\title{Recovery Lake Stage Envelope Score}
\usage{
rec_env(stg.data)
}
\arguments{
\item{stg.data}{see details}
}
\value{
Returns a \code{data.frame} of original data and normal stage elevation score.
}
\description{
Recovery Lake Stage Envelope Score
}
\details{
The input \code{stg.data} is a \code{data.frame} with columns:
\itemize{
\item{\code{Date} (as a \code{POSIXct} or \code{Date} variable)}
\item{\code{Data.Value} as stage elevation data in feet (NGVD29)}
}
}
\examples{

# Example dataset (not real data)
dates=seq(as.Date("2016-01-01"),as.Date("2016-02-02"),"1 days")
dat=data.frame(Date=dates,Data.Value=runif(33,12,18))

rec_env(dat)
}
